fix: 修复 flutter analyze 报错（CI lint 门禁失败）
- easing.dart: clamp() 返回 num 需 toDouble()，修复 double 返回类型不匹配
- zoom_curve.dart / zoom_curve_painter.dart / export_screen.dart: 移除未使用导入

+++ lib/core/math/easing.dart
@@ -14,7 +14,7 @@ double evaluateEasing(
   Offset control1 = const Offset(0.25, 0.1),
   Offset control2 = const Offset(0.25, 1.0),
 }) {
-  final u = t.clamp(0.0, 1.0);
+  final u = t.clamp(0.0, 1.0).toDouble();
   switch (type) {
     case EasingType.linear:
       return u;
